Add login with azaas form

+++ src/main/kotlin/com/example/services/AuthenticationService.kt
@@ -1,7 +1,10 @@
 package com.example.services
 
 import com.example.data.Token
-import com.example.helpers.*
+import com.example.helpers.APIConnector
+import com.example.helpers.SessionHelper
+import com.example.helpers.TOKEN_KEY
+import com.example.helpers.getToken
 import org.springframework.session.MapSessionRepository
 
 class AuthenticationService(
@@ -10,16 +13,16 @@ class AuthenticationService(
     private val sessionRepository: MapSessionRepository
 ) {
 
-    fun login(login: String, password: String) {
-        val session = sessionHelper.getSession()
-        session.setAttribute(TOKEN_KEY, Token(apiConnector.authenticateUser(login, password)))
-        session.setAttribute(LOGIN_KEY, login)
+    fun login(token: String) {
+        sessionHelper.getSession().setAttribute(TOKEN_KEY, Token(token))
     }
 
     fun logout() {
-        apiConnector.invalidateToken(sessionHelper.getSession().getToken()!!.value)
-        sessionHelper.getSession().removeAttribute(TOKEN_KEY)
-        sessionHelper.getSession().removeAttribute(LOGIN_KEY)
-        sessionRepository.deleteById(sessionHelper.getSession().id)
+        val token = sessionHelper.getSession().getToken()
+        if (token != null) {
+            apiConnector.invalidateToken(token.value)
+            sessionHelper.getSession().removeAttribute(TOKEN_KEY)
+            sessionRepository.deleteById(sessionHelper.getSession().id)
+        }
     }
 }
